% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadOFDS.R
\name{downloadOFDS}
\alias{downloadOFDS}
\title{Download OFDS data from Google drive}
\usage{
downloadOFDS(country, overwrite = FALSE)
}
\arguments{
\item{country}{A country folder on Google Drive - Currently must be 'Ghana' or 'Kenya'}

\item{overwrite}{logical - overwirte existing data?}
}
\value{
a list of data frames
}
\description{
Download OFDS data from Google drive
}
\examples{
\dontrun{
downloadOFDS('Ghana')
}

}
