% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapOFDS.R
\name{mapOFDS}
\alias{mapOFDS}
\title{Title}
\usage{
mapOFDS(
  ofdslist,
  spancol = "networkname",
  nodecol = "networkname",
  colpal = paletteer_dynamic("cartography::multi.pal", 20),
  spanlegend = TRUE,
  nodelegend = TRUE
)
}
\arguments{
\item{ofdslist}{a list of OFDS objects generates from readOFDS}

\item{spancol}{the variable to colour spans by}

\item{nodecol}{the variable to colour nodes by}

\item{colpal}{a colour palette for node and span colours}

\item{spanlegend}{logical - show legend?}

\item{nodelegend}{logical - show legend?}
}
\value{
a leaflet map
}
\description{
Title
}
\examples{
\dontrun{
dflist <- readOFDS()
mapOFDS(dflist,
spancol = 'networkname',
nodecol = 'networkname',
nodelegend = FALSE,
spanlegend = FALSE)}
}
