% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOFDS.R
\name{readOFDS}
\alias{readOFDS}
\title{Read in OFDS JSON data}
\usage{
readOFDS(dir)
}
\arguments{
\item{dir}{A url to a Google drive directory containing OFDS JSON objects}
}
\value{
a list of data frames
}
\description{
Read in OFDS JSON data
}
\examples{
\dontrun{
downloadOFDS('exampleURL.com')
dflist <- readOFDS('data/Kenya')
}


}
