% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOFDS.R
\name{readOFDS}
\alias{readOFDS}
\title{Title}
\usage{
readOFDS(dir)
}
\arguments{
\item{dir}{A directory containing OFDS JSON objects}
}
\value{
a list of data frames
}
\description{
Title
}
\examples{
\dontrun{
dflist <- mapOFDS(dd,
spancol = 'networkname',
nodecol = 'networkname',
nodelegend = FALSE,
spanlegend = FALSE)
dflist
}

}
